Initial work to check for suspensions in mngmt cmd
- Updates the twitterbot management command to make a request to check
against individual accounts to see if they are suspended
- Attempt to create a LoggedAction for this, and set the suspension on
the extra_data field on the PersonIdentifier

Setup Twitterbot check for suspended accounts

+++ ynr/apps/twitterbot/management/commands/twitterbot_update_usernames.py
@@ -1,7 +1,9 @@
+import requests
+from django.conf import settings
 from django.core.management.base import BaseCommand
 from django.db import transaction
-
 from people.models import Person
+from candidates.models import LoggedAction
 from twitterbot.helpers import TwitterBot
 
 from ..twitter import TwitterAPIData
@@ -33,7 +35,6 @@ class Command(BaseCommand):
         for identifier in twitter_identifiers:
             screen_name = identifier.value or None
             user_id = identifier.internal_identifier
-
             if user_id:
                 verbose(
                     "{person} has a Twitter user ID: {user_id}".format(
@@ -41,23 +42,56 @@ class Command(BaseCommand):
                     )
                 )
                 if user_id not in self.twitter_data.user_id_to_screen_name:
-                    print(
-                        "Removing user ID {user_id} for {person_name} as it is not a valid Twitter user ID. {person_url}".format(
-                            user_id=user_id,
-                            person_name=person.name,
-                            person_url=person.get_absolute_url(),
-                        )
+                    token = settings.TWITTER_APP_ONLY_BEARER_TOKEN
+                    headers = {
+                        "Authorization": "Bearer {token}".format(token=token)
+                    }
+                    r = requests.get(
+                        "https://api.twitter.com/1.1/users/show.json",
+                        params={"screen_name": screen_name},
+                        headers=headers,
                     )
-                    self.twitterbot.save(
-                        person,
-                        msg="This Twitter user ID no longer exists; removing it ",
-                    )
-                    if screen_name:
-                        self.twitterbot.save(
-                            person,
-                            msg="This Twitter screen name no longer exists; removing it ",
-                        )
-                    identifier.delete()
+                    data = r.json()
+
+                    if data:
+                        if any(d["code"] == 63 for d in data["errors"]):
+                            verbose(
+                                f"{person}'s Twitter account (id: {user_id}) is currently suspended."
+                            )
+                            if identifier.extra_data == {
+                                "status": "Twitter account (id: {user_id}) is currently suspended."
+                            }:
+                                continue
+                            else:
+                                identifier.extra_data = {
+                                    "status": "Twitter account (id: {user_id}) is currently suspended"
+                                }
+                                identifier.save()
+                                LoggedAction.objects.create(
+                                    user=self.user,
+                                    person=self.person,
+                                    source="TwitterBot",
+                                    action_type="suspended-twitter-account",
+                                )
+                                return identifier.extra_data
+                        else:
+                            print(
+                                "Removing user ID {user_id} for {person_name} as it is not a valid Twitter user ID. {person_url}".format(
+                                    user_id=user_id,
+                                    person_name=person.name,
+                                    person_url=person.get_absolute_url(),
+                                )
+                            )
+                            self.twitterbot.save(
+                                person,
+                                msg="This Twitter user ID no longer exists; removing it ",
+                            )
+                            if screen_name:
+                                self.twitterbot.save(
+                                    person,
+                                    msg="This Twitter screen name no longer exists; removing it ",
+                                )
+                            identifier.delete()
 
                     return
                 correct_screen_name = self.twitter_data.user_id_to_screen_name[
@@ -94,6 +128,22 @@ class Command(BaseCommand):
                     screen_name.lower()
                     not in self.twitter_data.screen_name_to_user_id
                 ):
+                    token = settings.TWITTER_APP_ONLY_BEARER_TOKEN
+                    headers = {
+                        "Authorization": "Bearer {token}".format(token=token)
+                    }
+                    r = requests.get(
+                        "https://api.twitter.com/1.1/users/show.json",
+                        params={"screen_name": screen_name},
+                        headers=headers,
+                    )
+                    data = r.json()
+
+                    if data:
+                        if any(d["code"] == 63 for d in data["errors"]):
+                            verbose(
+                                f"{person}'s Twitter account (id: {user_id}) is currently suspended."
+                            )
                     print(
                         "Removing screen name {screen_name} for {person_name} as it is not a valid Twitter screen name. {person_url}".format(
                             screen_name=screen_name,
